Separate UI/Algorithm into different modules
