--- modules/TakesStatistics/src/main/resources/nl/liacs/takesstatistics/Bundle.properties
@@ -1,17 +1,0 @@
-# To change this license header, choose License Headers in Project Properties.
-# To change this template file, choose Tools | Templates
-# and open the template in the editor.
-
-DiameterMetricPanelTopComponent.jXHeader1.description=Description and stuffs
-DiameterMetricPanelTopComponent.jXHeader1.title=Diameter & Radius Computation
-DiameterMetricPanelTopComponent.jRadioButton1.text=directed
-DiameterMetricPanelTopComponent.jRadioButton2.text=undirected
-DiameterMetricPanel.header.title=Diameter & Radius Computation
-# To change this license header, choose License Headers in Project Properties.
-# To change this template file, choose Tools | Templates
-# and open the template in the editor.
-DiameterMetricPanel.header.description=Calculate Diameter and Radius of the largest weakest connected component using the BoundingDiameter Algorithm.
-DiameterMetricPanel.eccentricitiesCheckbox.toolTipText=
-DiameterMetricPanel.eccentricitiesCheckbox.text=Compute all node eccentricities
-DiameterMetricPanel.peripheryCheckbox.text=Compute the periphery
-DiameterMetricPanel.centerCheckbox.text=Compute the center
