+++ modules/TakesStatisticsSuite/src/main/resources/nl/liacs/takesstatisticssuite/Bundle.properties
@@ -1,0 +1,6 @@
+#Localized module labels. Defaults taken from POM (<name>, <description>, <groupId>) if unset.
+#OpenIDE-Module-Name=
+#OpenIDE-Module-Short-Description=
+#OpenIDE-Module-Long-Description=
+#OpenIDE-Module-Display-Category=
+#Thu Oct 25 17:24:46 CEST 2018
