+++ modules/TakesStatisticsUI/pom.xml
@@ -1,25 +1,33 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    
+
     <parent>
         <artifactId>gephi-plugin-parent</artifactId>
         <groupId>org.gephi</groupId>
         <version>0.9.2</version>
     </parent>
-    
+
     <groupId>nl.liacs</groupId>
-    <artifactId>takesstatistics</artifactId>
+    <artifactId>takesstatisticsui</artifactId>
     <version>0.0.1</version>
     <packaging>nbm</packaging>
 
-    <name>Takes Statistics</name>
+    <name>Takes Statistics UI</name>
 
     <dependencies>
         <!-- Insert dependencies here -->
+        
+        <dependency>
+            <groupId>nl.liacs</groupId>
+            <artifactId>takesstatistics</artifactId>
+            <version>0.0.1</version>
+        </dependency>
+        
         <dependency>
             <groupId>org.netbeans.api</groupId>
             <artifactId>org-openide-util-lookup</artifactId>
+            <version>${netbeans.version}</version>
         </dependency>
         <dependency>
             <groupId>org.gephi</groupId>
@@ -29,15 +37,6 @@
         <dependency>
             <groupId>org.gephi</groupId>
             <artifactId>statistics-plugin</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.gephi</groupId>
-            <artifactId>utils-longtask</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.gephi</groupId>
-            <artifactId>graph-api</artifactId>
-            <type>jar</type>
         </dependency>
         <dependency>
             <groupId>org.swinglabs.swingx</groupId>
@@ -76,6 +75,7 @@
         </dependency>
     </dependencies>
 
+    
     <build>
         <plugins>
             <plugin>
@@ -83,7 +83,6 @@
                 <artifactId>nbm-maven-plugin</artifactId>
                 <configuration>
                     <publicPackages>
-                        nl.liacs.takesstatistics
                     </publicPackages>
                 </configuration>
             </plugin>
@@ -97,6 +96,7 @@
             </plugin>
         </plugins>
     </build>
+    
 </project>
