+++ modules/TakesStatisticsUI/src/main/java/nl/liacs/takesstatisticsui/DiameterMetricPanel.form
@@ -47,21 +47,21 @@
     <Component class="org.jdesktop.swingx.JXHeader" name="header">
       <Properties>
         <Property name="description" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.header.description" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.header.description" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
         <Property name="doubleBuffered" type="boolean" value="false"/>
         <Property name="title" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.header.title" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.header.title" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
     </Component>
     <Component class="javax.swing.JCheckBox" name="eccentricitiesCheckbox">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.eccentricitiesCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.eccentricitiesCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
         <Property name="toolTipText" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.eccentricitiesCheckbox.toolTipText" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.eccentricitiesCheckbox.toolTipText" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
       <Events>
@@ -71,7 +71,7 @@
     <Component class="javax.swing.JCheckBox" name="peripheryCheckbox">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.peripheryCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.peripheryCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
       <Events>
@@ -81,7 +81,7 @@
     <Component class="javax.swing.JCheckBox" name="centerCheckbox">
       <Properties>
         <Property name="text" type="java.lang.String" editor="org.netbeans.modules.i18n.form.FormI18nStringEditor">
-          <ResourceString bundle="nl/liacs/takesstatistics/Bundle.properties" key="DiameterMetricPanel.centerCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
+          <ResourceString bundle="nl/liacs/takesstatisticsui/Bundle.properties" key="DiameterMetricPanel.centerCheckbox.text" replaceFormat="org.openide.util.NbBundle.getMessage({sourceFileName}.class, &quot;{key}&quot;)"/>
         </Property>
       </Properties>
     </Component>
