+++ modules/TakesStatisticsUI/src/main/nbm/manifest.mf
