+++ modules/pom.xml
@@ -620,4 +620,7 @@
             </build>
         </profile>
     </profiles>
+    <modules>
+        <module>TakesStatisticsSuite</module>
+    </modules>
 </project>
