+++ pom.xml
@@ -11,8 +11,9 @@
     
     <!-- List of modules -->
     <modules>
-        <!-- Add here the paths of all modules (e.g. <module>modules/MyModule</module>) -->
+        <module>modules/TakesStatisticsSuite</module>
         <module>modules/TakesStatistics</module>
+        <module>modules/TakesStatisticsUI</module>
     </modules>
     
     <!-- Properties -->
